Make `ArbitraryOrderLagrangeElement` VTK-compatible
