+++ src/felupe/mesh/_geometry.py
@@ -18,6 +18,7 @@ along with FElupe.  If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 
+from ..element._lagrange import lagrange_hexahedron, lagrange_quad
 from ._line_rectangle_cube import cube_hexa, line_line, rectangle_quad
 from ._mesh import Mesh
 from ._tools import concatenate
@@ -281,152 +282,30 @@ class Grid(Mesh):
         )
 
 
-class RectangleArbitraryOrderQuad(Mesh):
-    """A rectangular 2d-mesh with arbitrary order quads between ``a`` and ``b`` with
-    ``n`` points per axis."""
+class RectangleArbitraryOrderQuad(Rectangle):
+    """A rectangular 2d-mesh with an arbitrarr-order Lagrange quad between ``a`` and
+    ``b``.
+    """
 
     def __init__(self, a=(0.0, 0.0), b=(1.0, 1.0), order=2):
-        yv, xv = np.meshgrid(
-            np.linspace(a[1], b[1], order + 1),
-            np.linspace(a[0], b[0], order + 1),
-            indexing="ij",
+        super().__init__(a=a, b=b, n=order + 1)
+        self.update(
+            cells=lagrange_quad(order=order).reshape(1, -1),
+            cell_type="VTK_LAGRANGE_QUADRILATERAL",
         )
 
-        points = np.vstack((xv.flatten(), yv.flatten())).T
 
-        # search vertices
-        xmin = min(points[:, 0])
-        ymin = min(points[:, 1])
-
-        xmax = max(points[:, 0])
-        ymax = max(points[:, 1])
-
-        def search_vertice(p, x, y):
-            return np.where(np.logical_and(p[:, 0] == x, p[:, 1] == y))[0][0]
-
-        def search_edge(p, a, x):
-            return np.where(p[:, a] == x)[0][1:-1]
-
-        v1 = search_vertice(points, xmin, ymin)
-        v2 = search_vertice(points, xmax, ymin)
-        v3 = search_vertice(points, xmax, ymax)
-        v4 = search_vertice(points, xmin, ymax)
-
-        vertices = [v1, v2, v3, v4]
-
-        mask1 = np.ones_like(points[:, 0], dtype=bool)
-        mask1[vertices] = 0
-        # points_no_verts = points[mask1]
-
-        e1 = search_edge(points, 1, ymin)
-        e2 = search_edge(points, 0, xmax)
-        e3 = search_edge(points, 1, ymax)
-        e4 = search_edge(points, 0, xmin)
-
-        edges = np.hstack((e1, e2, e3, e4))
-
-        mask2 = np.ones_like(points[:, 0], dtype=bool)
-        mask2[np.hstack((vertices, edges))] = 0
-        # points_no_verts_edges = points[mask2]
-
-        face = np.arange(len(points))[mask2]
-
-        cells = np.hstack((vertices, edges, face)).reshape(1, -1)
-
-        super().__init__(points, cells, cell_type="VTK_LAGRANGE_QUADRILATERAL")
-
-
-class CubeArbitraryOrderHexahedron(Mesh):
-    """A cube shaped 3d-mesh with arbitrary order hexahedrons between ``a`` and ``b``
-    with ``n`` points per axis."""
+class CubeArbitraryOrderHexahedron(Cube):
+    """A rectangular 2d-mesh with an arbitrarr-order Lagrange hexahedron between ``a``
+    and ``b``.
+    """
 
     def __init__(self, a=(0.0, 0.0, 0.0), b=(1.0, 1.0, 1.0), order=2):
-        zv, yv, xv = np.meshgrid(
-            np.linspace(a[2], b[2], order + 1),
-            np.linspace(a[1], b[1], order + 1),
-            np.linspace(a[0], b[0], order + 1),
-            indexing="ij",
+        super().__init__(a=a, b=b, n=order + 1)
+        self.update(
+            cells=lagrange_hexahedron(order=order).reshape(1, -1),
+            cell_type="VTK_LAGRANGE_HEXAHEDRON",
         )
-
-        points = np.vstack((xv.flatten(), yv.flatten(), zv.flatten())).T
-
-        # search vertices
-        xmin = min(points[:, 0])
-        ymin = min(points[:, 1])
-        zmin = min(points[:, 2])
-
-        xmax = max(points[:, 0])
-        ymax = max(points[:, 1])
-        zmax = max(points[:, 2])
-
-        def search_vertice(p, x, y, z):
-            return np.where(
-                np.logical_and(np.logical_and(p[:, 0] == x, p[:, 1] == y), p[:, 2] == z)
-            )[0][0]
-
-        def search_edge(p, a, b, x, y):
-            return np.where(np.logical_and(p[:, a] == x, p[:, b] == y))[0][1:-1]
-
-        def search_face(p, a, x, vertices, edges):
-            face = np.where(points[:, a] == x)[0]
-            mask = np.zeros_like(p[:, 0], dtype=bool)
-            mask[face] = 1
-            mask[np.hstack((vertices, edges))] = 0
-            return np.arange(len(p[:, 0]))[mask]
-
-        v1 = search_vertice(points, xmin, ymin, zmin)
-        v2 = search_vertice(points, xmax, ymin, zmin)
-        v3 = search_vertice(points, xmax, ymax, zmin)
-        v4 = search_vertice(points, xmin, ymax, zmin)
-
-        v5 = search_vertice(points, xmin, ymin, zmax)
-        v6 = search_vertice(points, xmax, ymin, zmax)
-        v7 = search_vertice(points, xmax, ymax, zmax)
-        v8 = search_vertice(points, xmin, ymax, zmax)
-
-        vertices = [v1, v2, v3, v4, v5, v6, v7, v8]
-
-        mask1 = np.ones_like(points[:, 0], dtype=bool)
-        mask1[vertices] = 0
-        # points_no_verts = points[mask1]
-
-        e1 = search_edge(points, 1, 2, ymin, zmin)
-        e2 = search_edge(points, 0, 2, xmax, zmin)
-        e3 = search_edge(points, 1, 2, ymax, zmin)
-        e4 = search_edge(points, 0, 2, xmin, zmin)
-
-        e5 = search_edge(points, 1, 2, ymin, zmax)
-        e6 = search_edge(points, 0, 2, xmax, zmax)
-        e7 = search_edge(points, 1, 2, ymax, zmax)
-        e8 = search_edge(points, 0, 2, xmin, zmax)
-
-        e9 = search_edge(points, 0, 1, xmin, ymin)
-        e10 = search_edge(points, 0, 1, xmax, ymin)
-        e11 = search_edge(points, 0, 1, xmax, ymax)
-        e12 = search_edge(points, 0, 1, xmin, ymax)
-        
-        edges = np.hstack((e1, e2, e3, e4, e5, e6, e7, e8, e9, e10, e11, e12))
-
-        mask2 = np.ones_like(points[:, 0], dtype=bool)
-        mask2[np.hstack((vertices, edges))] = 0
-        # points_no_verts_edges = points[mask2]
-
-        f1 = search_face(points, 0, xmin, vertices, edges)
-        f2 = search_face(points, 0, xmax, vertices, edges)
-        f3 = search_face(points, 1, ymin, vertices, edges)
-        f4 = search_face(points, 1, ymax, vertices, edges)
-        f5 = search_face(points, 2, zmin, vertices, edges)
-        f6 = search_face(points, 2, zmax, vertices, edges)
-
-        faces = np.hstack((f1, f2, f3, f4, f5, f6))
-
-        mask3 = np.ones_like(points[:, 0], dtype=bool)
-        mask3[np.hstack((vertices, edges, faces))] = 0
-        volume = np.arange(len(points))[mask3]
-
-        cells = np.hstack((vertices, edges, faces, volume)).reshape(1, -1)
-
-        super().__init__(points, cells, cell_type="VTK_LAGRANGE_HEXAHEDRON")
 
 
 class Circle(Mesh):
